+++ DeBroglie.Test/EdgedPathConstraintTest.cs
@@ -1,0 +1,128 @@
+﻿using DeBroglie.Constraints;
+using DeBroglie.Models;
+using DeBroglie.Topo;
+using NUnit.Framework;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace DeBroglie.Test
+{
+    [TestFixture]
+    class EdgedPathConstraintTest
+    {
+
+        // Test cloned from PathConstraintTest
+        // EdgedPath should behave identically if set up correctly.
+        [Test]
+        public void TestEdgedPathConstraint()
+        {
+            var a = new int[,]{
+                 {0, 0, 0, 1, 0, 0},
+                 {0, 0, 0, 1, 0, 0},
+                 {0, 0, 0, 1, 0, 0},
+                 {1, 1, 1, 1, 0, 0},
+                 {0, 0, 0, 0, 0, 0},
+                 {0, 0, 0, 0, 0, 0}
+            };
+
+            var allDirections = Enumerable.Range(0, 4).ToHashSet();
+
+            var exits = new Dictionary<Tile, ISet<int>>()
+            {
+                { new Tile(1), allDirections },
+            };
+
+            var model = OverlappingModel.Create(a, 3, false, 8);
+            var propagator = new TilePropagator(model, new Topology(10, 10, false), true, constraints: new[] {
+                new EdgedPathConstraint(exits, new []{new Point(0,0), new Point(9, 9) })
+            });
+            var status = propagator.Run();
+            Assert.AreEqual(Resolution.Decided, status);
+            var result = propagator.ToValueArray<int>().ToArray2d();
+            // Write out result for debugging
+            for (var y = 0; y < 10; y++)
+            {
+                for (var x = 0; x < 10; x++)
+                {
+                    Console.Write(result[x, y]);
+                }
+                Console.WriteLine();
+            }
+            // Simple flood fill algorithm to determine we have in fact got a path
+            var stack = new Stack<ValueTuple<int, int>>();
+            var visited = new bool[10, 10];
+            stack.Push((0, 0));
+            while (stack.TryPop(out var current))
+            {
+                var (x, y) = current;
+                if (x < 0 || x >= 10 || y < 0 || y >= 10)
+                    continue;
+                if (visited[x, y])
+                    continue;
+                visited[x, y] = true;
+                if (result[x, y] == 1)
+                {
+                    if (x == 9 && y == 9)
+                        return;
+                    stack.Push((x + 1, y));
+                    stack.Push((x - 1, y));
+                    stack.Push((x, y + 1));
+                    stack.Push((x, y - 1));
+                }
+            }
+            Assert.Fail();
+        }
+
+        [Test]
+        public void TestDirectionality()
+        {
+            var model = new AdjacentModel(Directions.Cartesian2d);
+            model.AddAdjacency(new Tile(1), new Tile(2), 1, 0, 0);
+
+            var topology = new Topology(2, 1, false);
+
+            var up = topology.Directions.GetDirection(0, 1);
+            var down = topology.Directions.GetDirection(0, -1);
+
+            var edgedPathConstraint = new EdgedPathConstraint(
+                new Dictionary<Tile, ISet<int>>()
+                {
+                   { new Tile(1), new[]{ up, down }.ToHashSet() },
+                   { new Tile(2), new[]{ up, down }.ToHashSet() },
+                }
+                );
+
+            var propagator = new TilePropagator(model, topology, constraints: new[] { edgedPathConstraint });
+
+            propagator.Run();
+
+            Assert.AreEqual(Resolution.Contradiction, propagator.Status);
+        }
+
+        [Test]
+        public void TestDirectionality2()
+        {
+            var model = new AdjacentModel(Directions.Cartesian2d);
+            model.AddAdjacency(new Tile(1), new Tile(2), 1, 0, 0);
+
+            var topology = new Topology(2, 1, false);
+
+            var left = topology.Directions.GetDirection(1, 0);
+            var right = topology.Directions.GetDirection(-1, 0);
+
+            var edgedPathConstraint = new EdgedPathConstraint(
+                new Dictionary<Tile, ISet<int>>()
+                {
+                   { new Tile(1), new[]{ left, right }.ToHashSet() },
+                   { new Tile(2), new[]{ left, right }.ToHashSet() },
+                }
+                );
+
+            var propagator = new TilePropagator(model, topology, constraints: new[] { edgedPathConstraint });
+
+            propagator.Run();
+        }
+    }
+}