+++ DeBroglie/Constraints/PathConstraintUtils.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using DeBroglie.Topo;
 
 namespace DeBroglie.Constraints
@@ -8,14 +9,48 @@ namespace DeBroglie.Constraints
     /// </summary>
     public static class PathConstraintUtils
     {
+        private static readonly int[] Emtpy = { };
+
+        public static SimpleGraph CreateGraph(Topology topology)
+        {
+            var nodeCount = topology.IndexCount;
+            var neighbours = new int[nodeCount][];
+            for (int i = 0; i < nodeCount; i++)
+            {
+                if(!topology.ContainsIndex(i))
+                {
+                    neighbours[i] = Emtpy;
+                }
+
+                var n = new List<int>();
+                for(var d=0;d<topology.Directions.Count;d++)
+                {
+                    if(topology.TryMove(i, d, out var dest))
+                    {
+                        n.Add(dest);
+                    }
+                }
+                neighbours[i] = n.ToArray();
+            }
+
+            return new SimpleGraph
+            {
+                NodeCount = nodeCount,
+                Neighbours = neighbours,
+            };
+        }
+
         /// <summary>
         /// Find articulation points.
         /// For an explanation, see:
         /// https://www.boristhebrave.com/2018/04/28/random-paths-via-chiseling/
         /// </summary>
-        public static bool[] GetArticulationPoints(Topology topology, bool[] walkable, bool[] relevant)
+        public static bool[] GetArticulationPoints(SimpleGraph graph, bool[] walkable, bool[] relevant)
         {
             var indices = walkable.Length;
+
+            if (indices != graph.NodeCount)
+                throw new Exception($"Length of walkable doesn't match count of nodes");
 
             var low = new int[indices];
             var num = 1;
@@ -33,15 +68,8 @@ namespace DeBroglie.Constraints
                 var isRelevantSubtree = isRelevant;
                 low[u] = dfsNum[u] = num++;
 
-                int ux, uy, uz;
-                topology.GetCoord(u, out ux, out uy, out uz);
-                for (var d = 0; d < topology.Directions.Count; d++)
+                foreach(var v in graph.Neighbours[u])
                 {
-                    int v;
-                    if (!topology.TryMove(ux, uy, uz, d, out v))
-                    {
-                        continue;
-                    }
                     if (!walkable[v])
                     {
                         continue;
@@ -95,6 +123,14 @@ namespace DeBroglie.Constraints
             }
 
             return isArticulation;
+        }
+
+
+        public class SimpleGraph
+        {
+            public int NodeCount { get; set; }
+
+            public int[][] Neighbours { get; set; }
         }
     }
 }